Refactor task_routes and cleaned up from the last wave

+++ app/routes/task_routes.py
@@ -1,9 +1,11 @@
-from flask import Blueprint, abort, make_response, request, Response
+import requests
+from flask import Blueprint, abort, make_response, request
 from ..db import db
 from app.models.task import Task
 from datetime import datetime
 import os
 from dotenv import load_dotenv
+from .route_utilities import validate_model
 
 load_dotenv()
 
@@ -14,13 +16,10 @@ def create_task():
     request_body = request.get_json()
     if "title" not in request_body or "description" not in request_body:
         abort(make_response({"details": "Invalid data"}, 400))
-    
 
-    title = request_body["title"]
-    description = request_body["description"]
     completed_at = request_body.get("completed_at", None)
-
-    new_task = Task(title=title, description=description, completed_at=completed_at)
+        
+    new_task = Task.from_dict(request_body)
     db.session.add(new_task)
     db.session.commit()
 
@@ -48,7 +47,9 @@ def get_all_tasks():
 
 @tasks_bp.get("/<task_id>")
 def get_one_task(task_id):
-    task = validate_task(task_id)
+    task = validate_model(Task, task_id)
+    if not task:
+        abort(make_response({"message": f"task_id {task_id} not found"}, 404))
     task_response = {
         "task": task.to_dict()  
     }
@@ -57,7 +58,9 @@ def get_one_task(task_id):
 
 @tasks_bp.put("/<task_id>")
 def update_task(task_id):
-    task = validate_task(task_id)
+    task = validate_model(Task, task_id)
+    if not task:
+        abort(make_response({"message": f"task_id {task_id} not found"}, 404))
     request_body = request.get_json()
 
     task.title = request_body["title"]
@@ -74,7 +77,7 @@ def update_task(task_id):
 
 @tasks_bp.patch("/<task_id>/mark_complete")
 def complete_task(task_id):
-    task = validate_task(task_id)
+    task = validate_model(Task, task_id)
 
     task.completed_at = datetime.now()  
     db.session.commit()
@@ -86,29 +89,32 @@ def complete_task(task_id):
         "Content-Type": "application/json"
     }
     request_body = {
-        "channell": "C07US894WRL",
+        "channel": "C07US894WRL",
         "text": f"Someone just completed the task {task.title}"
     }
-    notification = request.post(url, params=request_body, headers=headers)
+    response = requests.post(url, json=request_body, headers=headers)
 
-    if notification:
-        return {"task": task.to_dict()}, 200
+    if response.status_code == 200:
+        return{"task": task.to_dict()}, 200
+    else:
+        return {"error": "Failed to send Slack notification"}, 500
+    
 
 @tasks_bp.patch("/<task_id>/mark_incomplete")
 def mark_task_incomplete(task_id):
-    task = validate_task(task_id)
+    task = validate_model(Task, task_id)
+
     task.completed_at = None  
     db.session.commit()
 
-    response = {
-        "task": task.to_dict()
-    }
-    return response, 200
+    return {"task": task.to_dict()}, 200
 
 
 @tasks_bp.delete("/<task_id>")
 def delete_task(task_id):
-    task = validate_task(task_id)
+    task = validate_model(Task, task_id)
+    if not task:
+        abort(make_response({"message": f"task_id {task_id} not found"}, 404))
     
     db.session.delete(task)
     db.session.commit()
@@ -119,18 +125,5 @@ def delete_task(task_id):
 
     return response, 200
 
-
-def validate_task(task_id):
-    try:
-        task_id = int(task_id)
-    except ValueError:
-        abort(make_response({"message": f"task_id {task_id} invalid"}, 400))
-    
-    query = db.select(Task).where(Task.id == task_id)
-    task = db.session.scalar(query)
-
-    if not task:
-        abort(make_response({"message": f"task_id {task_id} not found"}, 404))
-    return task
